Retain the cursor position
Moving the cursor to the end of the line on each promote/demote is very
disturbing. It is better to stay on the same character.

This works for promoting and demoting as long as the number of
characters are not changed. But at the moment fails if the bullet is
replaced with another one with a different number of characters.
In that case we would have to recalculate the position based on the
changed characters.